Distance computation code refactored
& reorganized package structure, disovling datatype.tree package into main datatype package
